OpenRewrite 8.0 upgrade
