+++ src/main/java/org/openrewrite/java/micronaut/TypeRequiresIntrospection.java
@@ -15,22 +15,21 @@
  */
 package org.openrewrite.java.micronaut;
 
+import lombok.Data;
 import org.openrewrite.ExecutionContext;
-import org.openrewrite.Recipe;
-import org.openrewrite.SourceFile;
-import org.openrewrite.internal.ListUtils;
+import org.openrewrite.ScanningRecipe;
+import org.openrewrite.Tree;
+import org.openrewrite.TreeVisitor;
 import org.openrewrite.internal.lang.Nullable;
 import org.openrewrite.java.AnnotationMatcher;
 import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.java.JavaParser;
 import org.openrewrite.java.JavaTemplate;
-import org.openrewrite.java.search.UsesType;
 import org.openrewrite.java.tree.*;
 
-import java.time.Duration;
 import java.util.*;
 
-public class TypeRequiresIntrospection extends Recipe {
+public class TypeRequiresIntrospection extends ScanningRecipe<TypeRequiresIntrospection.Accumulator> {
     private static final Collection<String> typesRequiringIntrospection = Arrays.asList("io.micronaut.http.annotation.Controller", "io.micronaut.http.client.annotation.Client");
 
     @Override
@@ -41,11 +40,6 @@ public class TypeRequiresIntrospection extends Recipe {
     @Override
     public String getDescription() {
         return "In Micronaut 2.x a reflection-based strategy was used to retrieve that information if the class was not annotated with `@Introspected`. As of Micronaut 3.x it is required to annotate classes with `@Introspected` that are used in this way.";
-    }
-
-    @Override
-    public Duration getEstimatedEffortPerOccurrence() {
-        return Duration.ofMinutes(5);
     }
 
     private static boolean parentRequiresIntrospection(@Nullable JavaType.FullyQualified type) {
@@ -61,61 +55,51 @@ public class TypeRequiresIntrospection extends Recipe {
     }
 
     @Override
-    protected List<SourceFile> visit(List<SourceFile> before, ExecutionContext ctx) {
-        // look for classes requiring Introspected types
-        Set<JavaType.FullyQualified> typesFromSources = new HashSet<>();
-        for (SourceFile sourceFile : before) {
-            if (sourceFile instanceof J.CompilationUnit) {
-                J.CompilationUnit cu = (J.CompilationUnit) sourceFile;
-                for (J.ClassDeclaration classDeclaration : cu.getClasses()) {
-                    typesFromSources.add(classDeclaration.getType());
-                }
-            }
-        }
-
-        Set<JavaType.FullyQualified> introspectableTypes = new HashSet<>();
-
-        FindParamsAndReturnTypes findParamsAndReturnTypes = new FindParamsAndReturnTypes(typesFromSources);
-        for (SourceFile sourceFile : before) {
-            if (sourceFile instanceof J.CompilationUnit) {
-                J.CompilationUnit cu = (J.CompilationUnit) sourceFile;
-                for (J.ClassDeclaration classDeclaration : cu.getClasses()) {
-                    if (parentRequiresIntrospection(classDeclaration.getType())) {
-                        Set<JavaType.FullyQualified> paramAndReturnTypes = new HashSet<>();
-                        findParamsAndReturnTypes.visit(classDeclaration, paramAndReturnTypes);
-                        introspectableTypes.addAll(paramAndReturnTypes);
-                    }
-                }
-            }
-        }
-
-        return ListUtils.map(before, sourceFile -> {
-            if (sourceFile instanceof J.CompilationUnit) {
-                J.CompilationUnit cu = (J.CompilationUnit) sourceFile;
-                for (J.ClassDeclaration aClass : cu.getClasses()) {
-                    if (introspectableTypes.contains(aClass.getType())) {
-                        return (SourceFile) new AddIntrospectionAnnotationVisitor().visit(cu, introspectableTypes);
-                    }
-                }
-            }
-            return sourceFile;
-        });
+    public Accumulator getInitialValue() {
+        return new Accumulator();
     }
 
     @Override
-    protected JavaIsoVisitor<ExecutionContext> getApplicableTest() {
-        return new UsesType<>("io.micronaut..*", false);
+    public TreeVisitor<?, ExecutionContext> getScanner(Accumulator acc) {
+        return new TreeVisitor<Tree, ExecutionContext>() {
+            @Override
+            public @Nullable Tree visit(@Nullable Tree tree, ExecutionContext ctx) {
+                // look for classes requiring Introspected types
+                FindParamsAndReturnTypes findParamsAndReturnTypes = new FindParamsAndReturnTypes();
+                if (tree instanceof J.CompilationUnit) {
+                    J.CompilationUnit cu = (J.CompilationUnit) tree;
+                    for (J.ClassDeclaration classDeclaration : cu.getClasses()) {
+                        if (parentRequiresIntrospection(classDeclaration.getType())) {
+                            findParamsAndReturnTypes.visit(classDeclaration, acc.getIntrospectableTypes());
+                        }
+                    }
+                }
+                return tree;
+            }
+        };
+    }
+
+    @Override
+    public TreeVisitor<?, ExecutionContext> getVisitor(Accumulator acc) {
+        return new TreeVisitor<Tree, ExecutionContext>() {
+            @Override
+            public @Nullable Tree visit(@Nullable Tree tree, ExecutionContext ctx) {
+                if (tree instanceof J.CompilationUnit) {
+                    J.CompilationUnit cu = (J.CompilationUnit) tree;
+                    for (J.ClassDeclaration aClass : cu.getClasses()) {
+                        if (acc.getIntrospectableTypes().contains(aClass.getType())) {
+                            return new AddIntrospectionAnnotationVisitor().visit(cu, acc.getIntrospectableTypes());
+                        }
+                    }
+                }
+                return tree;
+            }
+        };
     }
 
     private static final class FindParamsAndReturnTypes extends JavaIsoVisitor<Set<JavaType.FullyQualified>> {
-        private final Set<JavaType.FullyQualified> typesInSourceSet;
-
-        private FindParamsAndReturnTypes(Set<JavaType.FullyQualified> typesInSourceSet) {
-            this.typesInSourceSet = typesInSourceSet;
-        }
-
         private void maybeAddType(@Nullable JavaType.FullyQualified type, Set<JavaType.FullyQualified> foundTypes) {
-            if (type != null && typesInSourceSet.contains(type) && !TypeRequiresIntrospection.parentRequiresIntrospection(type)) {
+            if (type != null && !TypeRequiresIntrospection.parentRequiresIntrospection(type)) {
                 foundTypes.add(type);
             }
         }
@@ -161,7 +145,7 @@ public class TypeRequiresIntrospection extends Recipe {
         final AnnotationMatcher INTROSPECTION_ANNOTATION_MATCHER = new AnnotationMatcher("@" + introspectedAnnotationFqn);
         final JavaTemplate templ = JavaTemplate.builder(this::getCursor, "@Introspected")
                 .imports(introspectedAnnotationFqn)
-                .javaParser(() -> JavaParser.fromJavaVersion().dependsOn("package io.micronaut.core.annotation; public @interface Introspected {}").build())
+                .javaParser(JavaParser.fromJavaVersion().dependsOn("package io.micronaut.core.annotation; public @interface Introspected {}"))
                 .build();
 
         @Override
@@ -177,5 +161,10 @@ public class TypeRequiresIntrospection extends Recipe {
             }
             return cd;
         }
+    }
+
+    @Data
+    static class Accumulator {
+        Set<JavaType.FullyQualified> introspectableTypes = new HashSet<>();
     }
 }
