+++ src/test/java/.editorconfig
@@ -1,0 +1,5 @@
+root = true
+
+[*.java]
+indent_size = 4
+ij_continuation_indent_size = 2
