--- src/test/kotlin/org/openrewrite/java/micronaut/UpgradeMicronautGradlePropertiesVersionTest.kt
@@ -1,38 +1,0 @@
-/*
- * Copyright 2021 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * https://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.openrewrite.java.micronaut
-
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.io.TempDir
-import org.openrewrite.properties.PropertiesRecipeTest
-import java.nio.file.Path
-
-@Suppress("UnusedProperty")
-class UpgradeMicronautGradlePropertiesVersionTest : PropertiesRecipeTest {
-
-    override val recipe =
-        UpgradeMicronautGradlePropertiesVersion("~2.1")
-
-    @Test
-    fun changeValue(@TempDir tempDir: Path) {
-        val properties = tempDir.resolve("gradle.properties").apply {
-            toFile().parentFile.mkdirs()
-            toFile().writeText("micronautVersion=2.0.3")
-        }.toFile()
-
-        assertChanged(recipe = recipe, before = properties, after = "micronautVersion=2.1.4")
-    }
-}
