--- src/test/kotlin/org/openrewrite/java/micronaut/UpgradeMicronautMavenPropertyVersionTest.kt
@@ -1,53 +1,0 @@
-/*
- * Copyright 2021 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * https://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.openrewrite.java.micronaut
-
-import org.junit.jupiter.api.Test
-import org.openrewrite.maven.Assertions.pomXml
-import org.openrewrite.test.RecipeSpec
-import org.openrewrite.test.RewriteTest
-
-class UpgradeMicronautMavenPropertyVersionTest : RewriteTest {
-    override fun defaults(spec: RecipeSpec) {
-        spec.recipe(UpgradeMicronautMavenPropertyVersion("~2.1"))
-    }
-
-    @Test
-    fun changeMavenMicronautVersion() = rewriteRun(
-        pomXml("""
-            <project>
-                <modelVersion>4.0.0</modelVersion>
-                <groupId>com.mycompany.app</groupId>
-                <artifactId>my-app</artifactId>
-                <version>1</version>
-                <properties>
-                    <micronaut.version>2.0.3</micronaut.version>
-                </properties>
-            </project>
-        """,
-        """
-            <project>
-                <modelVersion>4.0.0</modelVersion>
-                <groupId>com.mycompany.app</groupId>
-                <artifactId>my-app</artifactId>
-                <version>1</version>
-                <properties>
-                    <micronaut.version>2.1.4</micronaut.version>
-                </properties>
-            </project>
-        """)
-    )
-}
